[fix] move some components
