+++ src/components/Header/Header.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { NavLink } from "react-router-dom"
 
-import MenuButton from '../MenuButton'
+import HeaderMenuButton from '../HeaderMenuButton'
 import Logo from '../../assets/apple.png';
 import Search from '../Search'
 
@@ -21,8 +21,8 @@ const Header = (): React.ReactElement => {
           <Search />
         </div>
         <div className={styles.menu}>
-          <MenuButton className={styles.menuButton} href="/login" text="Login"/>
-          <MenuButton className={styles.menuButton} href="/id0" text="MyPage"/>
+          <HeaderMenuButton className={styles.menuButton} href="/login" text="Login"/>
+          <HeaderMenuButton className={styles.menuButton} href="/id0" text="MyPage"/>
         </div>
       </div>
       <hr/>
