+++ src/components/HeaderMenuButton/HeaderMenuButton.tsx
@@ -7,7 +7,7 @@ type MenuButtonProps = {
   text: string;
 }
 
-const MenuButton = (props: MenuButtonProps): React.ReactElement => {
+const HeaderMenuButton = (props: MenuButtonProps): React.ReactElement => {
   const {className, href, text} = props;
 
   return (
@@ -17,4 +17,4 @@ const MenuButton = (props: MenuButtonProps): React.ReactElement => {
   );
 };
 
-export default MenuButton;
+export default HeaderMenuButton;
