+++ src/components/MenuButton/index.tsx
@@ -1,1 +1,1 @@
-export { default } from './MenuButton'
+export { default } from './ActionMenuButton'
