--- src/pages/myPage/components/Info/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './Info'
