--- src/pages/myPage/components/MenuButton/ActionMenuButton.module.pcss
@@ -1,21 +1,0 @@
-.action__button {
-    margin: 10px 10px;
-    border-radius: 12px;
-    font-weight: 800;
-    height: 24px;
-    border: none;
-    outline: none;
-    background: transparent;
-    text-align:center;
-    cursor: pointer;
-    box-shadow: 4px 4px 8px 0 rgba(0,0,0,0.3), -4px -4px 8px 0 rgba(255,255,255,0.83), inset 2px 2px 6px 0 rgba(0,0,0,0), inset -2px -2px 6px 0 rgba(255,255,255,0);
-    transition: 0.2s;
-    
-    &:hover {
-        box-shadow: 4px 4px 12px 0 rgba(0,0,0,0), -4px -4px 12px 0 rgba(255,255,255,0), inset 2px 2px 6px 0 rgba(0,0,0,0), inset -2px -2px 6px 0 rgba(255,255,255,0);
-    }
-
-    &:active {
-        box-shadow: 4px 4px 12px 0 rgba(0,0,0,0), -4px -4px 12px 0 rgba(255,255,255,0), inset 2px 2px 6px 0 rgba(0,0,0,0.5), inset -2px -2px 6px 0 rgba(255,255,255,0.83);
-    }
-}
