+++ src/pages/myPage/components/MyInfo/MyInfo.module.pcss
@@ -22,6 +22,7 @@
 }
 
 .editButton {
+    border-radius: 25%;
     padding: 0;
     cursor: pointer;
     background-color: rgba(0,0,0,0);
