+++ src/pages/myPage/components/MyInfo/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './MyInfo'
