+++ src/pages/myPage/components/myAction/index.tsx
@@ -1,1 +1,1 @@
-export { default } from './myAction'
+export { default } from './MyAction'
