--- src/pages/personPage/components/Avatar/Avatar.module.pcss
@@ -1,5 +1,0 @@
-.avatar {
-    padding: 20px;
-    height: 300px;
-    width: 185px;
-}
