--- src/pages/personPage/components/Avatar/Avatar.module.pcss.d.ts
@@ -1,4 +1,0 @@
-declare const styles: {
-  readonly avatar: string;
-};
-export = styles;
