--- src/pages/personPage/components/Avatar/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './Avatar'
