--- src/pages/personPage/components/MenuButton/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './ActionMenuButton'
