+++ src/pages/personPage/components/PersonAction/PersonAction.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import PersonButton from '../PersonButton'
+import ActionButton from '../../../../components/ActionButton'
 
 import styles from './PersonAction.module.pcss';
 
@@ -33,7 +33,7 @@ const PersonAction = ():React.ReactElement => {
         <ul className={styles.person__action}>
           {data.map((el, idx) => (
             <li key={idx}>
-              <PersonButton 
+              <ActionButton 
                 name={el.name} 
                 text={el.text} 
                 onClick={el.onClick}
