--- src/pages/personPage/components/PersonButton/ActionButton.module.pcss.d.ts
@@ -1,4 +1,0 @@
-declare const styles: {
-  readonly action__button: string;
-};
-export = styles;
