--- src/pages/personPage/components/PersonButton/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './ActionButton'
